requirements.txt
Django==3.1.2
djangorestframework==3.12.1
pytest==6.1.2
psycopg2==2.8.5
model_bakery==1.2.0

